refactor: comment out login and register routes, update AboutUs page with carousel
