+++ ParqueInfantil/frontend/src/App.jsx
@@ -21,8 +21,8 @@ const App = () => {
         <Router>
             <Routes>
                 <Route path="/" element={<Home />} />
-                <Route path="/login" element={<Login />} />
-                <Route path="/register" element={<Register />} />
+                {/* <Route path="/login" element={<Login />} /> */}
+                {/* <Route path="/register" element={<Register />} /> */}
                 <Route path="/about" element={<AboutUs />} />
                 <Route path="/contact" element={<ContactUs />} />
                 <Route path="/admin/list-entities" element={<ListEntities />} />
